LIME-1071 Update to CRI Lib 3.0.1
Removed ClientFactoryService, and switch to CRI Lib ClientProviderFactory.
Align with contructor changes in CRI Lib services.
Added netty-http client to the build exclusions and placed them in the each lambda sub project (to fully take effect)

Remove versions specified for the main Jackson dependencies and defer to the ones pull in from software.amazon.awssdk:bom.

There is a breakage in a later jackson version, which crashes in the sdk classes which expect the older versions.

Note : jackson-datatype-jsr310 and jackson-datatype-jdk8 are not in the aws sdk and are custom dependencies these version have been set at the aws pom versions to avoid a mismatch.

Version Changes
    Correct missing aws platform bom
    AWS SDK 2.20.162 -> 2.26.16
    New Aws Crt Http Client aligned with AWS SDK 2.26.16
    AWS Lambda Events 3.11.0 -> 3.11.6
    Nimbusds Oauth 11.2 -> 11.4
    Jackson 2.13.1 -> 2.15.2 (needs to track the aws sdk version of jackson)